requirements.txt
requirements
numpy
pytorch
matplotlib
swig
gymnasium
gymnasium[box2d]